agentx/requirements.txt
python-dotenv>=1.0.1
pydantic>=2.6.0
openai>=1.37.0
anthropic>=0.34.0
httpx>=0.27.0
playwright>=1.45.0
tenacity>=8.3.0
cryptography>=42.0.5
numpy>=1.26.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
